+++ NEWS.md
@@ -81,6 +81,7 @@ Red Panda C++ Version 2.27
   - enhancement: Restore line position after file is modified outside and reloaded.
   - fix: Caret on '('/',' in string/comment shouldn't invoke function info tips.
   - fix: Function name not correctly found if it and the '(' is not in one line;
+  - fix: Register names in the cpu info are not in correct order.
   
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.